email-service/requirements.txt
celery
redis
python-dotenv
jinja2
requests